Moduling part main.rs
